+++ src/html/index.html
@@ -16,7 +16,7 @@
     <title>SCOPin rock: polarizing microscope simulator</title>
 
     <meta name="thumbnail"
-        content="https://s3-ap-northeast-1.amazonaws.com/fumipo-theta-microscope/images/social_card_icon.png" />
+        content="https://s3-ap-northeast-1.amazonaws.com/fumipo-theta-microscope/images/social_card_thumbnail.png" />
     <meta name="description"
         content="偏光顕微鏡による岩石薄片観察をシミュレートするアプリです. This application simulates polarizing microscopic view of rock slices." />
     <meta property="og:title" content="SCOPin rock: polarizing microscope simulator" />
