--- vender/app.config.ts
@@ -1,16 +1,0 @@
-import { Footer } from "@vender/component/footer/custom/footer"
-import { Social } from "@vender/component/social/custom/social"
-import { AppLogo } from "@vender/component/app_logo/app_logo"
-import { navigationMessage, welcomeMessage, viewerContainerMessage, sampleListMessage } from "@vender/i18n/message"
-
-const AppConfig = {
-    Footer,
-    Social,
-    AppLogo,
-    navigationMessage,
-    welcomeMessage,
-    viewerContainerMessage,
-    sampleListMessage,
-}
-
-export default AppConfig
