+++ vender/component/footer/default/footer.tsx
@@ -1,5 +1,7 @@
 import React from "react"
 
-export const Footer: React.FC = () => {
+const Footer: React.FC = () => {
     return <></>
 }
+
+export default Footer
