+++ vender/component/social/default/social.tsx
@@ -1,5 +1,7 @@
 import React from "react"
 
-export const Social: React.FC = () => {
+const Social: React.FC = () => {
     return <></>
 }
+
+export default Social
