+++ vender/custom_components.ts
@@ -1,0 +1,17 @@
+/**
+ * You can switch which implementation to be used for the components below.
+ * - AppLogo
+ * - Footer
+ * - Social
+ */
+import AppLogo from "@vender/component/app_logo/custom/app_logo"
+import Footer from "@vender/component/footer/custom/footer"
+import Social from "@vender/component/social/custom/social"
+
+const CustomComponents = {
+    Footer,
+    Social,
+    AppLogo,
+}
+
+export default CustomComponents
