+++ vender/resource/images/readme.md
@@ -1,0 +1,22 @@
+# About this directory
+
+This directory is for overriding images used in the application and images used in your custom components.
+The images here finally located in `release/images/`.
+
+You can override images below:
+
+- `favicon.png`
+  - used for "favicon" of this app
+- `logo.svg`
+  - used for open and crossed nicol toggle button
+  - rounded shape and the same width and height are recommended
+- `logo.png`
+  - possibly used for registering this app on the home screen of smart devices
+- `logo_192.png`
+  - possibly used for registering this app on the home screen of smart devices
+- `logo_512.png`
+  - possibly used for registering this app on the home screen of smart devices
+- `social_card_thumbnail`
+  - used for thumbnail of the social card when shared by a SNS service
+
+You can locate the other special images in this directory.
