+++ vender/resource/root/manifest.json
